# Conflicts:
#	SRSN/SRSN.Client_View/Pages/Index.cshtml
#	SRSN/SRSN.Client_View/Pages/Recipe_Detail.cshtml
#	SRSN/SRSN.Client_View/wwwroot/js/index.js
#	SRSN/SRSN.Client_View/wwwroot/js/recipe.js
#	SRSN/SRSN.Client_View/wwwroot/recipepress/css/custom.css
#	SRSN/SRSN.DatabaseManager/Entities/CookyDemoContext.cs

+++ SRSN/SRSN.DatabaseManager/Entities/RatingRecipe.cs
@@ -6,7 +6,7 @@ namespace SRSN.DatabaseManager.Entities
     public partial class RatingRecipe
     {
         public int Id { get; set; }
-        public double? Star { get; set; }
+        public int? Star { get; set; }
         public int? UserId { get; set; }
         public int? RecipeId { get; set; }
         public string ImageUrl { get; set; }
